feat fs: add search bar functionality
